third_party/WebKit/LayoutTests/fast/events/wheel/wheel-event-dispatch-event.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id=target></div>
<script type="text/javascript">
description('Test that wheel events should be dispatched and catched in the EventListener.');

var target = document.getElementById("target");
var wheelevent = new WheelEvent("wheelevent");
var recevied = false;
target.addEventListener("wheelevent", function() {
    recevied = true;
    debug("Wheelevent recevied");
});
target.dispatchEvent(wheelevent);

shouldBeTrue("recevied");
</script>
