third_party/blink/web_tests/fast/events/wheel/wheel-event-dispatch-event-expected.txt

Test that wheel events should be dispatched and catched in the EventListener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Wheelevent recevied
PASS recevied is true
PASS successfullyParsed is true

TEST COMPLETE
